refactor: move custom claims out to own project
