+++ .github/workflows/Build_and_Test.yml
@@ -4,7 +4,6 @@ on:
   push:
     branches:
       - main
-      - feat/dotnet9    
   pull_request:
     branches:
       - main
