+++ Serilog.Enrichers.AzureClaims.sln
@@ -14,7 +14,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Serilog.Enrichers.AzureClaims", "src\Serilog.Enrichers.AzureClaims\Serilog.Enrichers.AzureClaims.csproj", "{3FFC6CC2-072F-4E1A-8941-13030DC2049A}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Serilog.Enrichers.AzureClaims.Tests", "tests\Serilog.Enrichers.AzureClaims.Tests\Serilog.Enrichers.AzureClaims.Tests.csproj", "{EAECB584-06F0-47BD-8A95-479C7E51ACF2}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Serilog.Enrichers.Tests", "tests\Serilog.Enrichers.AzureClaims.Tests\Serilog.Enrichers.Tests.csproj", "{EAECB584-06F0-47BD-8A95-479C7E51ACF2}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Serilog.Enrichers.Claims", "src\Serilog.Enrichers.Claims\Serilog.Enrichers.Claims.csproj", "{448F70A7-C49B-42C0-896F-B8F01CD224AE}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -30,6 +32,10 @@ Global
 		{EAECB584-06F0-47BD-8A95-479C7E51ACF2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EAECB584-06F0-47BD-8A95-479C7E51ACF2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EAECB584-06F0-47BD-8A95-479C7E51ACF2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{448F70A7-C49B-42C0-896F-B8F01CD224AE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{448F70A7-C49B-42C0-896F-B8F01CD224AE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{448F70A7-C49B-42C0-896F-B8F01CD224AE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{448F70A7-C49B-42C0-896F-B8F01CD224AE}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
